Change Ticket definition for more flexibility
We leave it to the user, to get complete representations of the App and/or the
Revision. We can't assume presence or absence of the constructs as we can't reason
about the context. That should make it easy to leverage local copies/caches of
certain objects, but still gives enough information to construct the missing App
and/or revision in the case of absence.