Implement explicit time handling
The general premise is that high-level interfaces such as
App & Revision are now "values in time", specifically, they store
the revision at which they were created. This `rev` attribute is
never modified.

Updates to the coordinator state via these interfaces return *new*
instances, copies of the previous ones with a new revision number.

Given an Application *app* which is not registered, calling
*app.Register()* registers the application, and returns an interface
to this registered application, in other words, at the point in time
right after the application is registered.

    app2, err := app.Register()
    app.Unregister()  // Error: app isn't registered
    app2.Unregister() // Ok
    app2.Register()   // Error: app2 is already registered

    app3 = app2.FastForward(-1)
    app3.Register()   // Ok