Check if mercurial is working correctly and abort if not the case
Quite some users had failing mercurial installation. While this is
in theory out of scope for the visor installation it causes a lot
of trouble for users not being able to install visor (as a dependency
of bazooka-cli).

Waiting for feedback if the `hg --version` is enough to trigger the
problem as I can't recreate the problem here