Encapsulate Revision in serializable message.
We are adding lifecycle state to ``Revision`` for the purposes of the
Bazooka-hosted build pipeline, because the pipeline will create
packages via a reserved ``Application`` ``ProcType`` specification.
There is the possibility that the build process may fail, and thusly
we want a way to denote the finality of the candidate revision
builds.  The lifecycle states enable such a thing.

       [*]    [*]
        |      |
        ^      ^
        |      |
[0] -> [1] -> [2]
 |      |      |
 v      ^      ^
 |      |      |
[0] ->  /  ->  /

= 0: Proposed
A new revision is proposed and is thusly undergoing the build
process.  No artifact is assumed to exist.  This state is achieved
upon using ``Revision.Propose``.

= 1: Accepted
A new revision has been finalized for acceptance after the build
process has succeeded.  An artifact and state are set via
``Revision.Ratify``.

= 2: Rejected / Pending-Deletion
A revision has been reject and will be deleted, including its
artifacts.  This is achieved via ``Revision.Reject``.

= *: Terminus
The terminal state.