Do not emit start event without context
During the registratin of an instance the start path is set which
resulted in start events fired for instances which hadn't have an
ip or port yet. This break dependent code which expects the
carried Instance object to be fully populated. Due to that change
the event should be only expected when the Instance is actually
started.