Prevent instance stopping when not running
It used to be possible to force an invalid transition to the
'stopping' state from other states than 'running'. This is
considered harmful for correctnes of ordering as well as feedback
to users.

The change guards by always retrieving the latest representation
of an instance and validates that it's in the 'running' state and
proceeds only in this case. An error is returned indicating that
the current state is invalid for the intended operation.